requirements.txt
flask==2.3.2
google-api-python-client==2.86.0
python-dotenv==1.0.0
scikit-learn==1.2.2
numpy==1.24.3